requirements.txt
hyperliquid-python-sdk>=0.4.0
eth-account>=0.13.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
